requirements.txt
djangorestframework_simplejwt==4.7.1
django_heroku==0.3.1
Django==3.1.5
djangorestframework==3.12.4
python-dotenv==0.18.0
gunicorn==20.1.0
